integration: move test_connection_failure to new_session.rs
The test is about creating a new Session, so it fits there well.
